Fix content flex value and orgsnize links
The content flex value, being 1, did not center the content as intended. By removing the shrink value by making it zero, the content should enlargen to the center of the screen. Also, the margins of the links make them unorderly in their rows. The enclosed changes keep them in the corners of the screen.