Start final exercise
The enclosed changes are the starting changes for this final exercise. Organizing this correctly is paramount, and in order to fulfill the "holy grail" layout, it is simply a header, content, and footer box. The enclosed changes add some scaffolding to the html by grouping the sidebar and cards as "content". The css changes for the header have been included.